Refactor codebase
Refactor the existing codebase for improved readability and maintainability.
[skip gpt_engineer]
